requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Core ML
# ─────────────────────────────────────────────────────────────────────────────
torch>=2.1.0
transformers>=4.38.0
datasets>=2.18.0
accelerate>=0.28.0
sentencepiece>=0.1.99
timm>=0.9.16                # LayoutLMv3 vision backbone

# ─────────────────────────────────────────────────────────────────────────────
# Data & Image Processing
# ─────────────────────────────────────────────────────────────────────────────
opencv-python>=4.9.0.80
Pillow>=10.2.0
numpy>=1.26.0
pandas>=2.2.0
PyMuPDF>=1.23.0             # fitz — PDF loading

# ─────────────────────────────────────────────────────────────────────────────
# OCR Backends  (install at least one)
# ─────────────────────────────────────────────────────────────────────────────
paddlepaddle>=2.6.0         # PaddlePaddle CPU; use paddlepaddle-gpu for GPU
paddleocr>=2.7.0

# Tesseract fallback (optional — requires `brew install tesseract` on macOS)
pytesseract>=0.3.13

# ─────────────────────────────────────────────────────────────────────────────
# API
# ─────────────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9     # for file uploads

# ─────────────────────────────────────────────────────────────────────────────
# Config & Utilities
# ─────────────────────────────────────────────────────────────────────────────
pydantic>=2.6.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
tqdm>=4.66.0

# ─────────────────────────────────────────────────────────────────────────────
# MLOps (optional — comment out if not needed)
# ─────────────────────────────────────────────────────────────────────────────
mlflow>=2.11.0

# ─────────────────────────────────────────────────────────────────────────────
# Testing
# ─────────────────────────────────────────────────────────────────────────────
pytest>=8.1.0
pytest-cov>=5.0.0

# ─────────────────────────────────────────────────────────────────────────────
# Code Quality (optional — highly recommended)
# ─────────────────────────────────────────────────────────────────────────────
ruff>=0.3.0
